requirements.txt
#Django(LTS)
Django==2.2.9
#strictly separate the settings parameters
python-decouple
#safe db url,credentials
dj-database-url

#widget_tweaks
django-widget-tweaks

#Image upload
Pillow

#Rich editor
django-ckeditor
